Split catalog fetchers out of core.py into catalogs.py (#194)
Move apass_dr9, vsx_vizier and refcat2 from stellarphot.core into a new
stellarphot.catalogs module so core.py holds table data-structure classes
only. The fetchers stay importable from the top-level stellarphot namespace
(public, no warning) and from stellarphot.core via a deprecated PEP 562
__getattr__ shim that raises AstropyDeprecationWarning (removed in 3.0).

- catalogs.py imports CatalogData from core and PassbandMap from settings;
  the lazy shim import in core avoids a core <-> catalogs cycle.
- Update internal importers (utils/comparison_utils.py,
  utils/magnitude_transforms.py) to import from stellarphot.catalogs.
- Move the fetcher tests into tests/test_catalogs.py (structural shim tests
  plus the moved remote_data tests); CatalogData.from_vizier tests stay in
  test_core.py.
- Document stellarphot.catalogs (api_reference.rst, conf.py) and update the
  mermaid architecture diagrams (01, 02, 04) to show catalogs.py.
