+++ stellarphot/tests/test_core.py
@@ -18,9 +18,6 @@ from stellarphot.core import (
     CatalogData,
     PhotometryData,
     SourceListData,
-    apass_dr9,
-    refcat2,
-    vsx_vizier,
 )
 from stellarphot.settings import Camera, Observatory, PassbandMap
 
@@ -951,26 +948,6 @@ def test_catalog_to_passband_columns():
 
 
 @pytest.mark.remote_data
-def test_catalog_from_vizier_search_apass():
-    # Nothing special about this point...
-    sc = SkyCoord(ra=0, dec=0, unit="deg")
-
-    # Small enough radius to get only one star
-    radius = 0.03 * u.deg
-
-    # Use the APASS class factory to get the catalog
-    apass = apass_dr9(sc, radius=radius)
-    assert len(apass) == 6
-
-    # Calculated the value below by constructing a coordinate-based
-    # designation following IAU guidelines.
-    assert apass["id"][0] == "APASSSP J+359.9896+00.0122"
-
-    just_V = apass[apass["passband"] == "V"]
-    assert np.abs(just_V["mag"][0] - 15.559) < 1e-6
-
-
-@pytest.mark.remote_data
 @pytest.mark.parametrize("location_method", ["coord", "header", "wcs"])
 def test_catalog_from_vizier_search_vsx(location_method):
     # Do a cone search with a small enough radius to return exactly one star,
@@ -1064,205 +1041,10 @@ def test_from_vizier_with_header_no_wcs_raise_error():
         CatalogData.from_vizier(header, "B/vsx/vsx", clip_by_frame=True)
 
 
-@pytest.mark.remote_data
-@pytest.mark.parametrize(
-    "clip, data_file, mag_limit",
-    [
-        (True, "data/clipped_ey_uma_vsx.fits", None),
-        (False, "data/unclipped_ey_uma_vsx.fits", 13),
-    ],
-)
-def test_vsx_results(clip, data_file, mag_limit):
-    # Check that a catalog search of VSX gives us what we expect.
-    # I suppose this really isn't future-proof, since more variables
-    # could be discovered in the future....
-    data = get_pkg_data_filename(data_file)
-    expected = Table.read(data)
-    wcs_file = get_pkg_data_filename("data/sample_wcs_ey_uma.fits")
-    with fits.open(wcs_file) as hdulist:
-        with warnings.catch_warnings():
-            # Ignore the warning about the WCS having a different number of
-            # axes than the (non-existent) image.
-            warnings.filterwarnings(
-                "ignore",
-                message="The WCS transformation has more",
-                category=FITSFixedWarning,
-            )
-            wcs = WCS(hdulist[0].header)
-    CCD_SHAPE = [2048, 3073]
-    wcs.pixel_shape = list(reversed(CCD_SHAPE))
-    ccd = CCDData(data=np.zeros(CCD_SHAPE), wcs=wcs, unit="adu")
-
-    # Turn this into an HDU to get the standard FITS image keywords
-    ccd_im = ccd.to_hdu()
-
-    actual = vsx_vizier(
-        ccd_im[0].header,
-        radius=0.5 * u.degree,
-        clip_by_frame=clip,
-        magnitude_limit=mag_limit,
-    )
-
-    if mag_limit:
-        # If we have a magnitude limit, we need to filter the expected data
-        # to match.
-        expected = expected[expected["mag"] <= mag_limit]
-
-    assert set(actual["OID"]) == set(expected["OID"])
-
-
-@pytest.mark.remote_data
-@pytest.mark.parametrize(
-    "mag_limit,mag_limit_band",
-    [
-        (None, None),
-        (
-            13,
-            "V",
-        ),  # Limit chosen so that some of the expected data will be filtered out
-        (13, None),  # Default passband for apass_dr9 is V
-    ],
-)
-def test_find_apass(mag_limit, mag_limit_band):
-    CCD_SHAPE = [2048, 3073]
-    # This is really checking from APASS DR9 on Vizier, or at least that
-    # is where the "expected" data is drawn from.
-    expected_all = Table.read(get_pkg_data_filename("data/all_apass_ey_uma.ecsv"))
-
-    wcs_file = get_pkg_data_filename("data/sample_wcs_ey_uma.fits")
-    with fits.open(wcs_file) as hdulist:
-        with warnings.catch_warnings():
-            # Ignore the warning about the WCS having a different number of
-            # axes than the (non-existent) image.
-            warnings.filterwarnings(
-                "ignore",
-                message="The WCS transformation has more",
-                category=FITSFixedWarning,
-            )
-            wcs = WCS(hdulist[0].header)
-    wcs.pixel_shape = list(reversed(CCD_SHAPE))
-    ccd = CCDData(data=np.zeros(CCD_SHAPE), wcs=wcs, unit="adu")
-
-    # Turn this into an HDU to get the standard FITS image keywords
-    ccd_im = ccd.to_hdu()
-
-    all_apass = apass_dr9(
-        ccd_im[0].header,
-        radius=10 * u.arcmin,
-        magnitude_limit=mag_limit,
-        magnitude_limit_passband=mag_limit_band,
-    )
-
-    # Impose the magnitude limit on the expected result, if any
-    if mag_limit is not None:
-        expected_all = expected_all[expected_all["Vmag"] <= mag_limit]
-        # Apparently there are also some masked entries 🙄
-        expected_all = expected_all[~expected_all["Vmag"].mask]
-
-    # It is hard to imagine the RAs matching and other entries not matching,
-    # so just check the RAs.
-    assert set(ra.value for ra in all_apass["ra"]) == set(expected_all["RAJ2000"])
-
-    # The passbands ought to have been translated to the AAVSO standard names.
-    # This is a regression test for #439.
-    for band in ["B", "V", "SG", "SR", "SI"]:
-        assert band in all_apass["passband"]
-
-
-@pytest.mark.remote_data
-@pytest.mark.parametrize(
-    "mag_limit,mag_limit_band",
-    [
-        (None, None),
-        (
-            13,
-            "SR",
-        ),  # Limit chosen so that some of the expected data will be filtered out
-        (13, None),  # Default passband for refcat2 is SR
-    ],
-)
-def test_find_refcat2(mag_limit, mag_limit_band):
-    CCD_SHAPE = [2048, 3073]
-    # The "expected" data used for comparison is derived from refcat2 on Vizier.
-    expected_all = Table.read(get_pkg_data_filename("data/all_refcat2_ey_uma.ecsv"))
-
-    wcs_file = get_pkg_data_filename("data/sample_wcs_ey_uma.fits")
-    with fits.open(wcs_file) as hdulist:
-        with warnings.catch_warnings():
-            # Ignore the warning about the WCS having a different number of
-            # axes than the (non-existent) image.
-            warnings.filterwarnings(
-                "ignore",
-                message="The WCS transformation has more",
-                category=FITSFixedWarning,
-            )
-            wcs = WCS(hdulist[0].header)
-    wcs.pixel_shape = list(reversed(CCD_SHAPE))
-    ccd = CCDData(data=np.zeros(CCD_SHAPE), wcs=wcs, unit="adu")
-
-    # Turn this into an HDU to get the standard FITS image keywords
-    ccd_im = ccd.to_hdu()
-    all_refcat2 = refcat2(
-        ccd_im[0].header,
-        radius=10 * u.arcmin,
-        magnitude_limit=mag_limit,
-        magnitude_limit_passband=mag_limit_band,
-    )
-
-    # Impose the magnitude limit on the expected result, if any
-    if mag_limit is not None:
-        expected_all = expected_all[expected_all["rmag"] <= mag_limit]
-
-    # # It is hard to imagine the RAs matching and other entries not matching,
-    # # so just check the RAs.
-    assert set(ra.value for ra in all_refcat2["ra"]) == set(expected_all["RA_ICRS"])
-
-    # # The passbands ought to have been translated to the AAVSO standard names.
-    for band in ["GBP", "GRP", "GG", "SG", "SR", "SI", "SZ", "J", "H", "K"]:
-        assert band in all_refcat2["passband"]
-
-
 # Load test apertures
 test_sl_data = ascii.read(
     get_pkg_data_filename("data/test_sourcelist.ecsv"), format="ecsv", fast_reader=False
 )
-
-
-@pytest.mark.parametrize("catalog", [apass_dr9, refcat2, vsx_vizier])
-def test_catalog_errors(catalog):
-    # Check some of the errors expected in catalog classes
-
-    # Giving a bad band should raise an error
-    error_msg = (
-        "magnitude_limit_passband must be one of"
-        if catalog is not vsx_vizier
-        else "no straightforward way to limit the VSX catalog by passband"
-    )
-    with pytest.raises(ValueError, match=error_msg):
-        catalog(
-            {},  # Dummy header
-            radius=0.1 * u.arcmin,
-            magnitude_limit=13,
-            magnitude_limit_passband="not a band, more like an ensemble",
-        )
-
-    # Giving a band but not a magnitude limit should raise an error except
-    # for vsx_vizier, which has weird settings because the VSX catalog mag
-    # column has a bunch of different passbands in it.
-    if catalog is not vsx_vizier:
-        # Need to provide a valid passband for each catalog
-        passband = "SR"
-        with pytest.raises(ValueError, match="you provide a .* you must also provide"):
-            catalog(
-                {},  # Dummy header
-                radius=0.1 * u.arcmin,
-                magnitude_limit_passband=passband,
-            )
-
-    # Giving a magnitude limit but not a band should not raise an error because each
-    # catalog has a default passband.
-    #
-    # Since magnitude limit functionality is test elsewhere, we don't test it here.
 
 
 def test_sourcelist():
